Refactor HTTP server to hpp/cpp so it can be used in multiple TUs
